Move classes to make modules smaller and better named
