+++ pyiron_contrib/__init__.py
@@ -20,7 +20,7 @@ JOB_CLASS_DICT['ProtocolFTSEvolution'] = 'pyiron_contrib.protocol.compound.finit
 JOB_CLASS_DICT['ProtocolFTSEvolutionParallel'] = 'pyiron_contrib.protocol.compound.finite_temperature_string'
 JOB_CLASS_DICT['ImageJob'] = 'pyiron_contrib.image.job'
 JOB_CLASS_DICT['RandSpg'] = 'pyiron_contrib.atomistic.atomistics.structures.randspg'
-JOB_CLASS_DICT['Fenics'] = 'pyiron_contrib.continuum.fenics'
+JOB_CLASS_DICT['Fenics'] = 'pyiron_contrib.continuum.fenics.job'
 
 from ._version import get_versions
